added ordering

+++ assets/js/app/controllers/adminController.js
@@ -15,6 +15,7 @@ app.controller('AdminController', ['$http', '$scope',
 
 		me.mapCategory = function(c){
 			c.isExpanded = false;
+			c.games = c.games || [];
 			c.gameModel = {
 				points: 5,
 				difficulty: 'Easy',
@@ -54,6 +55,7 @@ app.controller('AdminController', ['$http', '$scope',
 				return;
 			}
 			me.loading = true;
+			me.model.order = me.categories.length;
 			$http.post('/api/category', me.model)
 			.success(function(data){
 				me.loadCategories();
@@ -73,7 +75,14 @@ app.controller('AdminController', ['$http', '$scope',
 				return;
 			}
 			me.loading = true;
-			$http.post('/api/game', cat.gameModel)
+			var query;
+			if(cat.gameModel.id){
+				query = $http.put('/api/game/' + cat.gameModel.id, cat.gameModel)
+			}else{
+				cat.gameModel.order =  cat.games.length;
+				query = $http.post('/api/game', cat.gameModel);
+			}
+			query
 			.success(function(data){
 				$http.get('/api/category/' + cat.id)
 				.success(function(data){
@@ -89,9 +98,26 @@ app.controller('AdminController', ['$http', '$scope',
 			});	
 		}
 
+		me.editGame = function(cat, game){
+			cat.gameModel = game;
+		}
+
+		me.cancelEditGame = function(cat, game){
+			cat.gameModel = {
+				points: 5,
+				difficulty: 'Easy',
+				category: cat.id,
+				author: me.user.id
+			};
+		}
+
 		me.deleteGame = function(game, index){
+			if(!confirm('Are you sure you want to delete the game and all its results?')){
+				return;
+			}
 			$http.delete('/api/game/' + game.id)
 			.success(function(data){
+				//ToDo: reorder after delete.
 				me.loadCategories();
 			})
 			.error(function(data){
@@ -99,6 +125,56 @@ app.controller('AdminController', ['$http', '$scope',
 				console.log(data);
 				alert('Error adding game.');
 			});		
+		}
+
+		me.move = function(arr, index, isCategory, isUp, event){
+			var mapModel = function(item){
+				var model = {};
+				for(var key in item){
+					if(item.hasOwnProperty(key)){
+						if( (key != 'games') && (key != 'id')){
+							model[key] = item[key];
+						}
+					}
+				}
+				return model;
+			}
+			if(me.loading){
+				return;
+			}
+			me.loading = true;
+			event.stopPropagation();
+			var entity, otherIndex;
+			if(isCategory){
+				entity = 'category';
+			}else{
+				entity = 'game';
+			}
+
+			if(isUp){
+				otherIndex = index - 1;
+			}else{
+				otherIndex = index + 1;
+			}
+
+			var tmp = arr[index].order || 0;
+			arr[index].order = arr[otherIndex].order;
+			arr[otherIndex].order = tmp;
+			$http.put('/api/' + entity + '/' + arr[index].id, mapModel(arr[index]))
+			.success(function(){
+				$http.put('/api/' + entity + '/' + arr[otherIndex].id, mapModel(arr[otherIndex]))
+				.success(function(){
+					me.loading = false;
+				})
+				.error(function(data){
+					console.log(data);
+					alert('Error');
+				});	
+			})
+			.error(function(data){
+				console.log(data);
+				alert('Error');
+			});
 		}
 	}]
 );
